Updated docker-compose tag
The tag for this dockerfile was set to `latest`; however, this tag does not exist. If you try to use this tag, you end up with a generic `manifest error`. 

The solution to this problem is to switch the tag to `main` instead. This gets the image to download correctly. 

Also updated the README to reflect this change.